feat(PlayerService): Split classes into multiple files
